屏幕圆角特性消除告警修改 Signed-off-by: panjiao<panjiao5@huawei.com>

+++ rosen/modules/render_service/core/pipeline/round_corner_display/rs_rcd_surface_render_node.cpp
@@ -143,8 +143,8 @@ void RSRcdSurfaceRenderNode::PrepareHardwareResourceBuffer(rs_rcd::RoundCornerLa
     }
     layerBitmap = *(layerInfo->curBitmap);
 
-    uint32_t bitmapHeight = layerBitmap.height();
-    uint32_t bitmapWidth = layerBitmap.width();
+    uint32_t bitmapHeight = static_cast<uint32_t>(layerBitmap.height());
+    uint32_t bitmapWidth = static_cast<uint32_t>(layerBitmap.width());
     SetRcdBufferHeight(bitmapHeight);
     SetRcdBufferWidth(bitmapWidth);
     SetRcdBufferSize(cldLayerInfo.bufferSize);
@@ -200,20 +200,20 @@ bool RSRcdSurfaceRenderNode::FillHardwareResource(HardwareLayerInfo &cldLayerInf
         uint32_t baseColor = 0;
     };
 
-    const int bytesPerPixel = 4; // 4 means four bytes per pixel
+    const uint32_t bytesPerPixel = 4; // 4 means four bytes per pixel
     CldInfo cldInfo;
-    cldInfo.cldSize = cldLayerInfo.bufferSize;
-    cldInfo.cldWidth = cldLayerInfo.cldWidth;
-    cldInfo.cldHeight = cldLayerInfo.cldHeight;
-    cldInfo.cldStride = cldLayerInfo.cldWidth * bytesPerPixel;
-    cldInfo.exWidth = width;
-    cldInfo.exHeight = height;
+    cldInfo.cldSize = static_cast<uint32_t>(cldLayerInfo.bufferSize);
+    cldInfo.cldWidth = static_cast<uint32_t>(cldLayerInfo.cldWidth);
+    cldInfo.cldHeight = static_cast<uint32_t>(cldLayerInfo.cldHeight);
+    cldInfo.cldStride = static_cast<uint32_t>(cldLayerInfo.cldWidth * bytesPerPixel);
+    cldInfo.exWidth = static_cast<uint32_t>(width);
+    cldInfo.exHeight = static_cast<uint32_t>(height);
     
     int offset = 0;
     int offsetCldInfo = 0;
     offsetCldInfo = height * stride;
     offset = (height + 1) * stride;
-    cldInfo.cldDataOffset = offset;
+    cldInfo.cldDataOffset = static_cast<uint32_t>(offset);
     
     errno_t ret = memcpy_s(reinterpret_cast<void*>(img + offsetCldInfo), sizeof(cldInfo), &cldInfo, sizeof(cldInfo));
     if (ret != EOK) {
